\name{bivarRIpower-package}
\alias{bivarRIpower}
\docType{package}
\title{
Sample size calculations for bivariate longitudinal data
}
\description{
Provides function to carry out sample size calculations for correlations that occur between random intercepts, residuals, and observations
in a bivariate longitudinal model.  Maximum likelihood estimates are used for correlation variances.  Sample size is calculated under
one-sample z-test framework.  Details of calculations are given in Comulada and Weiss (2010).
}
\details{
\tabular{ll}{
Package: \tab bivarRIpower\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2010-12-02\cr
License: \tab GPL (\>=2)\cr
LazyLoad: \tab yes\cr
}

}
\author{
W. Scott Comulada and Robert E. Weiss

Maintainer: W. Scott Comulada <scomulad@ucla.edu>

}
\references{
Comulada WS and Weiss RE. (2010). Power calculations for correlations between bivariate longitudinal data. Statistics in Medicine. 29(27): 2811-2824.
}

\keyword{ bivarRIpower }

